logo added to navbar
