+++ client/src/components/header/Header.js
@@ -1,0 +1,15 @@
+
+
+const Header = () => {
+  const [isOpen, setIsOpen] = React.useState(false)
+
+  const toggle = () => setIsOpen(!isOpen)
+
+  return(
+    ...
+   <MenuToggle toggle={toggle} isOpen={isOpen}>
+    ...
+  )
+}
+
+export default Header;
