+++ client/src/components/logo/Logo.js
@@ -10,3 +10,4 @@ export default function Logo(props) {
     </Box>
   )
 }
+
