+++ client/src/components/menuitem/MenuItem.js
@@ -1,0 +1,31 @@
+import React from "react"
+import { Link, Text, Stack, Box } from "@chakra-ui/react"
+
+const MenuItem = ({ children, isLast, to = "/", isOpen, ...rest }) => {
+  return (
+    <Link href={to}>
+      <Text display="block" {...rest}>
+        {children}
+      </Text>
+    </Link>
+  )
+}
+
+<Box
+  display={{ base: isOpen ? "block" : "none", md: "block" }}
+  flexBasis={{ base: "100%", md: "auto" }}
+>
+<Stack
+  spacing={8}
+  align="center"
+  justify={["center", "space-between", "flex-end", "flex-end"]}
+  direction={["column", "row", "row", "row"]}
+  pt={[4, 4, 0, 0]}
+>
+  <MenuItem to="/">Home</MenuItem>
+  <MenuItem to="/how">How It Works</MenuItem>
+  ...
+</Stack>
+</Box>
+
+export default MenuItem;
